Move Title implementation to a separate file
Re ECKIT-620
